+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+- Fix LSP completion for options.
+- Add LSP document highlighting support.
 - Add LSP completion for fully-qualified type references.
 
 ## [v1.63.0] - 2026-01-06